requirements.txt
appdirs==1.4.4
autopep8==1.5.7
black==21.7b0
click==8.0.1
Flask==2.0.1
Flask-Script==2.0.6
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.0
pathspec==0.8.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
regex==2021.7.6
six==1.16.0
toml==0.10.2
tomli==1.0.4
Werkzeug==2.0.1
